requirements.txt
youtube-transcript-api # for fetching YouTube video transcripts
langchain-community # for community-contributed LangChain integrations
langchain-openai # for OpenAI integrations with LangChain
faiss-cpu # for efficient similarity search and clustering of dense vectors
openai # for interacting with OpenAI's API
tiktoken # for tokenizing text for OpenAI models
python-dotenv # for loading environment variables from a .env file
langchain-core